services/ml-enrichment/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-aiplatform==1.*
google-cloud-logging==3.*
spacy==3.*
scikit-learn==1.*
numpy==1.*

# Enhanced ML features (optional but recommended)
sentence-transformers==2.*
transformers==4.*
torch==2.*
